requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
bcrypt
python-multipart
aiofiles
python-jose
fastapi-cache2
aioredis